+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+### Removed
+- Drop deprecated `pointcloud` and `heatmapgl` traces from the API
+
 ### Updated
 
 - Updated plotly.py to use base64 encoding of arrays in plotly JSON to improve performance.